+++ tests/wallet/test_account.py
@@ -1,13 +1,22 @@
 import unittest
-from neo3.wallet import Account, private_key_from_nep2, private_key_to_nep2
+
+from neo3.wallet import Account, private_key_to_nep2, private_key_from_nep2
 
 
 class AccountCreationTestCase(unittest.TestCase):
 
-    def test_createNep2Key(self):
+    def test_account_from_nep2(self):
+        nep2 = '6PYWFgCbuow5AwyfUXWyn1Cjc4NFFvrGq97nKoEXgvFiysHzstgt3ZT25G'
+        account = Account.from_nep2(nep2, "password")
+        print(account.address)
+        assert account.get_key() == nep2
+
+
+    def test_create_nep2(self):
         private_key = b'\x01' * 32
         password = "password"
         nep2 = private_key_to_nep2(private_key, password)
+        print(nep2)
         decrypted_key = private_key_from_nep2(nep2, password)
         assert private_key == decrypted_key
